+++ frontend/src/utils/buildGridTemplateColumns.js
